requirements.txt
torch
transformers
onnx
onnxruntime
numpy
opencv-python
Pillow
scipy